backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
gunicorn==21.2.0
email-validator==2.1.0